Rename polkadot network model file for consistency (#7695)
* chore: rename polkadot network model file


* chore: update readme


* fix: lint issue


---------
