+++ .changeset/odd-masks-sleep.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/coin-polkadot": patch
+---
+
+Rename network model types
