+++ libs/coin-modules/coin-polkadot/src/network/sidecar.ts
@@ -25,7 +25,7 @@ import type {
   SidecarRuntimeSpec,
   SidecarConstants,
   BlockInfo,
-} from "./sidecar.types";
+} from "./types";
 import { createRegistryAndExtrinsics } from "./common";
 import node from "./node";
