+++ libs/coin-modules/coin-polkadot/src/test/cli.ts
@@ -5,10 +5,7 @@ import flatMap from "lodash/flatMap";
 import { getCryptoCurrencyById } from "@ledgerhq/cryptoassets/index";
 import { isAccount } from "@ledgerhq/coin-framework/account";
 import { formatCurrencyUnit } from "@ledgerhq/coin-framework/currencies";
-import {
-  SidecarValidatorsParamAddresses,
-  SidecarValidatorsParamStatus,
-} from "../network/sidecar.types";
+import { SidecarValidatorsParamAddresses, SidecarValidatorsParamStatus } from "../network/types";
 import { AccountLike } from "@ledgerhq/types-live";
 import { PolkadotAccount, PolkadotValidator, Transaction } from "../types";
 import polkadotAPI from "../network";
